# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface Persona {
  id: string;
  name: string;
  description: string;
  gender: string;
  tagline: string;
}

export interface CompanionMessage {
  role: string;
  content: string;
}

export interface CompanionChatInput {
  sessionId: string;
  persona: string;
  messages: CompanionMessage[];
  voice?: boolean;
}

export interface CompanionReply {
  responseText: string;
  sessionId: string;
  /** @nullable */
  audioBase64?: string | null;
}

export interface CompanionMemory {
  sessionId: string;
  summary: string;
  /** @nullable */
  messageCount?: number | null;
  /** @nullable */
  updatedAt?: string | null;
}

export interface CompanionMemoryInput {
  sessionId: string;
  persona: string;
  messages: CompanionMessage[];
}