# lib/api-zod/src/generated/types/companionChatInput.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */
import type { CompanionMessage } from './companionMessage';

export interface CompanionChatInput {
  sessionId: string;
  persona: string;
  messages: CompanionMessage[];
  voice?: boolean;
}