# lib/api-zod/src/generated/types/companionMemory.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CompanionMemory {
  sessionId: string;
  summary: string;
  /** @nullable */
  messageCount?: number | null;
  /** @nullable */
  updatedAt?: string | null;
}