# lib/api-zod/src/generated/types/companionReply.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface CompanionReply {
  responseText: string;
  sessionId: string;
  /** @nullable */
  audioBase64?: string | null;
}