# lib/api-zod/src/generated/types/persona.ts
/**
 * Generated by orval v8.9.1 🍺
 * Do not edit manually.
 * Api
 * API specification
 * OpenAPI spec version: 0.1.0
 */

export interface Persona {
  id: string;
  name: string;
  description: string;
  gender: string;
  tagline: string;
}